+++ packages/gcs-store/src/index.ts
@@ -1,6 +1,6 @@
-import {Bucket} from '@google-cloud/storage'
+import type {Bucket, CreateWriteStreamOptions} from '@google-cloud/storage'
 import stream from 'node:stream'
-import http from 'node:http'
+import type http from 'node:http'
 import debug from 'debug'
 
 import {ERRORS, TUS_RESUMABLE, Upload, DataStore} from '@tus/utils'
@@ -37,13 +37,16 @@ export class GCSStore extends DataStore {
 
       file.storage = {type: 'gcs', path: file.id, bucket: this.bucket.name}
 
-      const options = {
+      const options: CreateWriteStreamOptions = {
         metadata: {
           metadata: {
             tus_version: TUS_RESUMABLE,
             ...this.#stringifyUploadKeys(file),
           },
         },
+      }
+      if (file.metadata?.contentType) {
+        options.contentType = file.metadata.contentType
       }
       const fake_stream = new stream.PassThrough()
       fake_stream.end()
@@ -133,7 +136,7 @@ export class GCSStore extends DataStore {
         return
       }
 
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      // biome-ignore lint/suspicious/noExplicitAny: todo
       this.bucket.file(id).getMetadata((error: any, metadata: any) => {
         if (error && error.code === 404) {
           return reject(ERRORS.FILE_NOT_FOUND)