+++ packages/utils/src/kvstores/RedisKvStore.ts
@@ -1,6 +1,6 @@
-import {RedisClientType} from '@redis/client'
-import {KvStore} from './Types'
-import {Upload} from '../models'
+import type {RedisClientType} from '@redis/client'
+import type {KvStore} from './Types'
+import type {Upload} from '../models'
 
 /**
  * Redis based configstore.
@@ -8,7 +8,10 @@ import {Upload} from '../models'
  * @author Mitja Puzigaća <mitjap@gmail.com>
  */
 export class RedisKvStore<T = Upload> implements KvStore<T> {
-  constructor(private redis: RedisClientType, private prefix: string = '') {
+  constructor(
+    private redis: RedisClientType,
+    private prefix = ''
+  ) {
     this.redis = redis
     this.prefix = prefix
   }
@@ -26,7 +29,14 @@ export class RedisKvStore<T = Upload> implements KvStore<T> {
   }
 
   async list(): Promise<Array<string>> {
-    return this.redis.keys(this.prefix + '*')
+    const keys = new Set<string>()
+    let cursor = 0
+    do {
+      const result = await this.redis.scan(cursor, {MATCH: `${this.prefix}*`, COUNT: 20})
+      cursor = result.cursor
+      for (const key of result.keys) keys.add(key)
+    } while (cursor !== 0)
+    return Array.from(keys)
   }
 
   private serializeValue(value: T): string {