+++ packages/utils/src/kvstores/index.ts
@@ -1,4 +1,5 @@
 export {FileKvStore} from './FileKvStore'
 export {MemoryKvStore} from './MemoryKvStore'
 export {RedisKvStore} from './RedisKvStore'
+export {IoRedisKvStore} from './IoRedisKvStore'
 export {KvStore} from './Types'